# specification-module/src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Store API - OpenAPI 3.0
  description: |-
    This is a Store API Server
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: https://store-api.swagger.io/api/v3
...
paths:
...
  /category:
    put:
      tags:
        - category
      summary: Update an existing Category
      description: Update an existing Category by Id
      operationId: updateCategory
      requestBody:
        description: Update an existent Category in the app
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/CategoryDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CategoryDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Category not found
        '405':
          description: Validation exception
      security:
        - bearerAuth: []
    post:
      tags:
        - category
      summary: Add a new Category to the app
      description: Add a new Category to the app
      operationId: addCategory
      requestBody:
        description: Create a new Category to the app
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryDto'
          application/xml:
            schema:
              $ref: '#/components/schemas/CategoryDto'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CategoryDto'
        required: true
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '405':
          description: Invalid input
      security:
        - bearerAuth: []
  /category/findByStatus:
    get:
      tags:
        - category
      summary: Finds Categorys by status
      description: Multiple status values can be provided with comma separated strings
      operationId: findCategoryByStatus
      parameters:
        - name: status
          in: query
          description: Status values that need to be considered for filter
          required: false
          explode: true
          schema:
            type: string
            default: ACTIVE
            enum:
              - ACTIVE
              - DEACTIVE
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponseDto'
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          description: Invalid status value
      security:
        - bearerAuth: []
  /category/{categoryId}:
    get:
      tags:
        - category
      summary: Find Category by ID
      description: Returns a single Category
      operationId: getCategoryById
      parameters:
        - name: categoryId
          in: path
          description: ID of Category to return
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
            application/xml:
              schema:
                $ref: '#/components/schemas/CategoryResponseDto'
        '400':
          description: Invalid ID supplied
        '404':
          description: Category not found
      security:
        - bearerAuth: []
    post:
      tags:
        - category
      summary: Updates a category in the app with form data
      description: ''
      operationId: updateCategoryWithForm
      parameters:
        - name: categoryId
          in: path
          description: ID of Parent Category to return
          required: true
          schema:
            type: integer
            format: int64
        - name: name
          in: query
          description: Name of category that needs to be updated
          schema:
            type: string
        - name: status
          in: query
          description: Status of category that needs to be updated
          schema:
            type: string
      responses:
        '405':
          description: Invalid input
      security:
        - bearerAuth: []
    delete:
      tags:
        - category
      summary: Deletes a category
      description: delete a category
      operationId: deleteCategory
      parameters:
        - name: api_key
          in: header
          description: ''
          required: false
          schema:
            type: string
        - name: categoryId
          in: path
          description: Category id to delete
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '400':
          description: Invalid category value
      security:
        - bearerAuth: []
...
components:
  schemas:
    ParentCategoryDto:
      required:
        - name
        - photoUrls
      type: object
      properties:
        parentCategoryId:
          type: integer
          format: int64
          example: 10
        name:
          type: string
          example: Lubricants
        photoUrl:
          type: string
          example: Lubricants
        status:
          type: string
          description: Parent Category status in the store
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: ParentCategoryDto
    CategoryDto:
      required:
        - categoryName
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
          example: 10
        categoryName:
          type: string
          example: Engine Oil, Break Oil
        photoUrl:
          type: string
        parentCategoryId:
          type: integer
          format: int64
          example: 10
        status:
          type: string
          description: Category status in the store
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: category
    CategoryResponseDto:
      required:
        - categoryName
      type: object
      properties:
        categoryId:
          type: integer
          format: int64
          example: 10
        categoryName:
          type: string
          example: Engine Oil, Break Oil
        photoUrl:
          type: string
        parentCategory:
          $ref: '#/components/schemas/ParentCategoryDto'
        status:
          type: string
          description: category status in the app
          enum:
            - ACTIVE
            - DEACTIVE
      xml:
        name: categoryResponse
...
  securitySchemes:
    bearerAuth:            # arbitrary name for the security scheme
      type: http
      scheme: bearer
      bearerFormat: JWT